Convert single-layer Louvain to Leiden
Also add simple shims and warnings for deprecated Louvain usage.
