+++ tests/test_champ_usage_3D.py
@@ -1,7 +1,7 @@
 from .shared_testing_functions import generate_connected_multilayer_ER, generate_random_values, \
     generate_random_partitions
 from modularitypruning.champ_utilities import CHAMP_3D
-from modularitypruning.louvain_utilities import multilayer_louvain_part_with_membership, \
+from modularitypruning.leiden_utilities import multilayer_louvain_part_with_membership, \
     check_multilayer_louvain_capabilities
 from numpy import mean
 from random import seed
